LPS-138553 Moves to service module, no logic changes
